xrRender: fix and improve screenshot capturing, misc font stuff
First of all, this started as "let's pull as many FreeType and
HarfBuzz dependencies as possible" in preparation for switching
font rendering to SFML::Text (now I have a different idea, see
below). So, here we are:

FreeType:
* brotli to support WOFF fonts. This actually might be handy;
* bzip2 to support bzip-compressed fonts (not sure I'll need it,
  but let it be just for the sake of completeness);
* png to support pngs embedded in fonts (happens sometimes);
* zlib[-ng] to support gzip-compressed fonts (and to build png).

HarfBuzz:
* graphite to support complex scripts.

Later I realized that some of these deps allow me to have some
fun. SquashFS now supports gzip compression, but I don't plan
to use it.

More important is how I now save screenshots. No WIC (Windows
Imaging Component) anymore. 7 formats to choose from. Support
for HDR render targets (some day I'll implement HDR, this will
come handy. In Anomaly modded exes, you can't save screenshots
when HDR is enabled, just because they don't have support for
any image format that handles HDR).
We have 3 types of screenshots: regular ones (F12), savegame
thumbnails and levelmap/cubemap. No idea what the last one is,
but I didn't remove it.

* regular ones: in SDR, you can now save screenshots in: bmp,
  dds, exr, jpg, png, qoi, tga. dds is handled by DirectXTex,
  exr by OpenEXR, png by its official library, the rest by
  SFML. dds screenshots are compressed on the GPU in BC7.
  png screenshots (default) are now saved quickly and with
  better compression. Thanks to the official lib and zlib-ng.
  All those claims WIC is the best are worth nothing.
  In HDR, only exr is available. dds supports HDR with BC6H,
  but it has no alpha channel. png supports 64 bpp, but in
  uint format which doesn't fit us.
* savegame thumbnails: dds for SDR. Instead of saving them in
  768x768 RGBA32, they're now saved in 1024x1024 BC7. Also,
  when shrinking, we use BasisU to perform resampling rather
  than DirectXTex, as the latter only has basic filters (and
  I like Kaiser). Again, for HDR only exr is available, but
  it's better than nothing. It's also filtered with Kaiser.
* levelmaps/cubemaps: tga for SDR, exr for HDR. Nothing
  interesting here.

I could break r__screenshot.cpp into smaller pieces, but it's
only 350 lines, I don't think it's worth it.

Memory usage when saving screenshots is reduced by manually
freeing the buffers as soon as they're not used anymore.

I also noticed that I broke savegame thumbnails some time ago.
I mean, they're still saved properly, but they capture the console
and/or the mainmenu is saving from it. It's related to coroutines
and the async queue, but I haven't found any other way to fix this
other than postponing the screenshot by 2 frames instead of 1.
Now everything is saved as it should be without any regressions.

Misc changes: pull OpCoDe updates from the ODE main repo. There
were some fixes for modern systems and compilers. It's been
working fine all that time, but let it be.

Now regarding the font rendering: the main problem with SFML is
that it doesn't handle line breaking/wrapping anyhow. We have our
own parser which does that, but it simply inserts a newline instead
of a space or a tab. I don't think it's a good/complete solution
when you have bidirectional text and Emoji support and all that.A

But I found Skribidi, library which does +/- the same that
SFML::Text, but also handles complex paragraphs and proper line
breaking. And Emojis/icons. So for now I'm targeting this library,
I don't quite understand how it works without FreeType (only
HarfBuzz), but I like how the final text looks like. Apart from
HarfBuzz and Sheenbidi, it would also require 2 more small
libraries, which is perfectly fine. The amount of code and effort
saved will be much bigger.

+++ ogsr_engine/xrCDB/xrCDB_box.cpp
@@ -269,7 +269,7 @@ public:
         return true;
     }
 
-    constexpr void _prim(u32 prim)
+    constexpr void _prim(size_t prim)
     {
         const auto id = gsl::narrow<s32>(prim);
         const TRI& T = tris[id];
