+++ ogsr_engine/xrCDB/xrCDB_ray.cpp
@@ -448,7 +448,7 @@ public:
         return true;
     }
 
-    constexpr void _prim(u32 prim)
+    constexpr void _prim(size_t prim)
     {
         const auto id = gsl::narrow<s32>(prim);
         const auto& tri = tris[id];
